Refactored route and loading logic
