+++ app/hooks/index.ts
@@ -3,3 +3,4 @@ export * from './useShareCredentials';
 export * from './usePendingCredential';
 export * from './useVerifyCredential';
 export * from './useRequestCredential';
+export * from './useAppLoading';
